refactor: move dupes payload contracts to api
